Added new features: - Tilt shift blur - Vignette - Image rotation
+ Fixes and minor changes

+++ src/libLMTKimage/imageshader.cpp
@@ -4,9 +4,9 @@ namespace image {
 
     void ImageShader::operator()(Image& img)
     {
-        for (size_t y = 0; y < img.height(); y++)
+        for (int y = 0; y < img.height(); y++)
         {
-            for (size_t x = 0; x < img.width(); x++)
+            for (int x = 0; x < img.width(); x++)
             {
                 img.setRGBAPixel((*this)(img, x, y), x, y);
             }
